+++ components/HabitList.tsx
@@ -41,17 +41,17 @@ export default function HabitList() {
 
   return (
     <div className="container mx-auto px-4 py-8">
-      <div className="flex justify-between items-center mb-6">
-        <h1 className="text-3xl font-bold">
-          {isTasksView ? 'My Tasks' : 'My Habits'}
-        </h1>
-        <Button onClick={() => setModalConfig({ isOpen: true, isTask: isTasksView })}>
-          <Plus className="mr-2 h-4 w-4" /> {isTasksView ? 'Add Task' : 'Add Habit'}
-        </Button>
-      </div>
-      <div className='py-4'>
-        <ViewToggle />
-      </div>
+             <div className="flex justify-between items-center mb-6">
+         <h1 className="text-3xl font-bold">
+           {isTasksView ? 'My Tasks' : 'My Habits'}
+         </h1>
+         <Button onClick={() => setModalConfig({ isOpen: true, isTask: isTasksView })}>
+           <Plus className="mr-2 h-4 w-4" /> {isTasksView ? 'Add Task' : 'Add Habit'}
+         </Button>
+       </div>
+       <div className='py-4'>
+         <ViewToggle />
+       </div>
       <div className="grid grid-cols-1 md:grid-cols-2 gap-4 items-stretch">
         {activeHabits.length === 0 ? (
           <div className="col-span-2">